requirements.txt
ultralytics==8.1.0
google-genai==0.1.0
opencv-python-headless==4.9.0.80
pandas==2.2.0
python-dotenv==1.0.0
fastapi
uvicorn[standard]
python-multipart
Flask
SQLAlchemy
PyMySQL
requests
